Interim checkin of changes to call intrinsic TeX commands and fragments with the keyboard. I am experimenting with using Ctrl+Space to start, and Return to end, the entry of a macro name. The only intrinsics defined so far are the lower case Greek letters up to omicron.

+++ prince/sidebar/content/sidebarOverlay.js
@@ -1487,6 +1487,117 @@ if (!SB_DEBUG) {
 addEventListener("load", sidebar_overlay_init, false);
 addEventListener("unload", sidebar_overlay_destruct, false);
 
+///////////////////////////////////////////
+// Beginning of code added by BBM. Much of the above code will be deleted
+//
+//
+
+
+function focusOnEditor()
+{
+  var editWindow = document.getElementById('content-frame').focus;
+}
+
+// we keep two objects that serve as associative arrays for the macros and fragments
+var macroArray;
+var fragmentArray;
+
+// builds the directory tree of fragment files and builds the list of fragment files for
+// calling by the keyboard. The boolean 'buildListOnly' is true when we don't need to rebuild
+// the tree but do need to build the list, as, for example, right after the ref has been
+// assigned.
+function rebuildFragmentTree(tree, buildListOnly)
+{
+  if (!buildListOnly) tree.builder.rebuild();
+  var ACSA = Components.classes["@mozilla.org/autocomplete/search;1?name=stringarray"].getService();
+  ACSA.QueryInterface(Components.interfaces.nsIAutoCompleteSearchStringArray);
+  var dir = Components.classes["@mozilla.org/file/local;1"].createInstance(Components.interfaces.nsILocalFile);
+  var dirpath = tree.getAttribute("ref");
+  dirpath = dirpath.substring(8,dirpath.length-1);
+    // take ref attribute and lop off "file:///" and last "/"
+  // for Windows
+  dirpath = dirpath.replace("/","\\","g");
+  dir.initWithPath(dirpath);
+   addFragmentsToList(ACSA, dir);
+}
+
+
+// A recursively-called routine that does the adding of the file names to the list
+function addFragmentsToList(autocomplete, dir)
+{
+  if (dir.isDirectory())
+  {
+    var enumerator = dir.directoryEntries.QueryInterface(Components.interfaces.nsIDirectoryEnumerator);
+    var file;
+    var dirpath = dir.path;
+    var filename;
+    while (enumerator.hasMoreElements())
+    {
+      file = enumerator.getNext();
+      file.QueryInterface(Components.interfaces.nsIFile);
+      if (file.isDirectory()) addFragmentsToList(autocomplete,file);
+      else
+      {
+        filename = file.leafName;
+        if (filename.length - filename.search(/\.frg$/)==4)
+        {
+          filename = filename.substring(0, filename.length-4);
+          autocomplete.addString("fragments",filename);
+          fragmentArray[filename] = new Object;
+          fragmentArray[filename].dirpath = dirpath;
+        }
+      }
+    }
+    enumerator.close();
+  }
+}  
+  
+
+function initSidebar()
+{
+  // Since RDF file trees can't take RESOURCE:// path names, we need to convert the
+  // fragmentsBaseDirectory to a FILE:// url.
+  macroArray = new Object();
+  fragmentArray = new Object();
+  var dsprops = Components.classes["@mozilla.org/file/directory_service;1"].getService(Components.interfaces.nsIProperties);
+  var basedir =dsprops.get("resource:app", Components.interfaces.nsIFile);
+  basedir.append("res");
+  var dir1 = basedir.clone();
+  dir1.append("fragments");
+  var dirpath = dir1.path + "/";
+  // The next line is for Windows only
+  dirpath = dirpath.replace("\\","/","g");
+  
+  var fragmentsBaseDirectory = "file:///" + dirpath;
+  document.getElementById("frag-tree").setAttribute("ref", fragmentsBaseDirectory);
+  rebuildFragmentTree(document.getElementById("frag-tree"),true);
+  //
+  // Now load the macros file
+  // We need to prebuild these so that the keyboard shortcut works
+  // ACSA = autocomplete string array
+  var ACSA = Components.classes["@mozilla.org/autocomplete/search;1?name=stringarray"].getService();
+  ACSA.QueryInterface(Components.interfaces.nsIAutoCompleteSearchStringArray);
+  var macrofile = basedir;
+  macrofile.append("tagdefs");
+  macrofile.append("macros.xml");
+  var xmlDoc = document.implementation.createDocument("", "macros", null);
+  var nodeList;
+  var s;
+  xmlDoc.async = false;
+  if (xmlDoc.load("file:///"+macrofile.path))
+  {
+    nodeList = xmlDoc.getElementsByTagName("m");
+    for (var i = 0; i < nodeList.length; i++)
+    {
+      s = nodeList.item(i).getAttribute("nm");
+      macroArray[s] = new Object();
+      macroArray[s].forcesMath = (nodeList.item(i).getAttribute("fm") =="1");
+      macroArray[s].data = nodeList.item(i).firstChild.textContent;
+      ACSA.addString("macros",s);
+    }
+  }
+}
+
 
 function loadFragment(event,tree)
 {
@@ -1535,20 +1646,51 @@ function loadFragment(event,tree)
         while (node && node.nodeType != node.CDATA_SECTION_NODE) node = node.nextSibling;
         if (node) infoString = node.nodeValue;
       }
-      var editorElement = document.getElementById("content-frame");
-      var editor = msiGetEditor(editorElement);
-      try
-      {
-        insertXMLAtCursor(editor, dataString, false, true);
-//      editor.insertHTML(dataString);
-      } catch(e) {}
+      insertDataAtCursor(dataString);
     }
   }
   catch(e)
   {
 //    alert(e);
   }
+  focusOnEditor();
 }
+
+function insertDataAtCursor( dataString )
+{
+  var editorElement = document.getElementById("content-frame");
+  var editor = msiGetEditor(editorElement);
+  try
+  {
+    insertXMLAtCursor(editor, dataString, false, true);
+  //      editor.insertHTML(dataString);
+  } catch(e) {}
+}
+
+
+
+function onMacroOrFragmentEntered( aString )
+{
+  var s = macroArray[aString];
+  if (s) 
+  {
+    if (s.forcesMath == true)
+    {
+      doParamCommand('cmd_MSIsymbolCmd',s.data);
+    }
+    else insertDataAtCursor( s.data );
+  }
+  else
+  {
+    s = fragmentArray[aString];
+    if (s) insertDataAtCursor( s.data );
+  }
+  var macrofragmentStatusPanel = document.getElementById('macroEntryPanel');
+  if (macrofragmentStatusPanel)
+    macrofragmentStatusPanel.setAttribute("hidden", "true");
+  focusOnEditor();
+}
+
 
 var fragObserver = 
 { 
@@ -1571,7 +1713,8 @@ var fragObserver =
     {
       var xmlDoc = document.implementation.createDocument("", "frag", null);
       xmlDoc.async = false;
-      if (xmlDoc.load(tree.getAttribute("ref") + "/" +s))
+      var path = tree.getAttribute("ref") + "/" + s;
+      if (xmlDoc.load(path))
       {
         var dataString=null;
         var node;
@@ -1584,12 +1727,14 @@ var fragObserver =
         }
         if (dataString.length == 0) return;
         transferData.data = new TransferData();
+        transferData.data.addDataForFlavour("privatefragmentfile", path);
         transferData.data.addDataForFlavour("text/html",dataString);
         transferData.data.addDataForFlavour("text/unicode",dataString);
         
       }
     }
     catch(e) {}
+    focusOnEditor();
   },
   
   onDrop: function(evt, dropData, session)
@@ -1617,44 +1762,58 @@ var fragObserver =
       }
     }
     dump("New fragment file path is "+pathbase + path + "\n");
-    var data = new Object;
-    window.openDialog("chrome://prince/content/fragmentname.xul", "", "modal,chrome,resizable=yes", data);
-    if (data.filename.length > 0)
+    if (dropData.flavour == "privatefragmentfile")
     {
-      var sFileContent = '<?xml version="1.0"?>\n<fragment>\n  <data>\n    <![CDATA[' +dropData.data+
-        ']]>\n  </data>\n  <context/>\n  <info/>\n</fragment>';
-      var filepath = pathbase + path + data.filename;
-      if (filepath.search(/.frg/) == -1) filepath += ".frg";
-      filepath = filepath.replace("/","\\","g");
-      try
+      var origPath = dropData.data;
+      // now move origPath to path
+      var origfile = Components.classes["@mozilla.org/file/local;1"].createInstance(Components.interfaces.nsILocalFile);
+      file.QueryInterface(Components.interfaces.nsIFile);
+      file.initWithPath( origPath );
+      file.moveTo(path);
+    }
+    else
+    {
+      var data = new Object;
+      window.openDialog("chrome://prince/content/fragmentname.xul", "", "modal,chrome,resizable=yes", data);
+      if (data.filename.length > 0)
       {
-        var file = Components.classes["@mozilla.org/file/local;1"].createInstance(Components.interfaces.nsILocalFile);
-        file.QueryInterface(Components.interfaces.nsIFile);
-        file.initWithPath( filepath );
-        if( file.exists() == true ) file.remove( false );
-        var strm = Components.classes["@mozilla.org/network/file-output-stream;1"].createInstance(Components.interfaces.nsIFileOutputStream);
-        strm.QueryInterface(Components.interfaces.nsIOutputStream);
-        strm.QueryInterface(Components.interfaces.nsISeekableStream);
-        strm.init( file, 0x04 | 0x08, 420, 0 );
-        strm.write( sFileContent, sFileContent.length );
-        strm.flush();
-        strm.close();
+        var sFileContent = '<?xml version="1.0"?>\n<fragment>\n  <data>\n    <![CDATA[' +dropData.data+
+          ']]>\n  </data>\n  <context/>\n  <info/>\n</fragment>';
+        var filepath = pathbase + path + data.filename;
+        if (filepath.search(/.frg/) == -1) filepath += ".frg";
+        filepath = filepath.replace("/","\\","g");
+        try
+        {
+          var file = Components.classes["@mozilla.org/file/local;1"].createInstance(Components.interfaces.nsILocalFile);
+          file.QueryInterface(Components.interfaces.nsIFile);
+          file.initWithPath( filepath );
+          if( file.exists() == true ) file.remove( false );
+          var strm = Components.classes["@mozilla.org/network/file-output-stream;1"].createInstance(Components.interfaces.nsIFileOutputStream);
+          strm.QueryInterface(Components.interfaces.nsIOutputStream);
+          strm.QueryInterface(Components.interfaces.nsISeekableStream);
+          strm.init( file, 0x04 | 0x08, 420, 0 );
+          strm.write( sFileContent, sFileContent.length );
+          strm.flush();
+          strm.close();
+        }
+        catch(ex)
+        {
+          window.alert(ex.message);
+        } 
+        rebuildFragmentTree(tree,false);
       }
-      catch(ex)
-      {
-        window.alert(ex.message);
-      } 
-      tree.setAttribute("ref",saveref);
     }
   },
   
   onDragOver: function(evt, flavour, session) 
   {
+    dump(flavour.contentType+"\n");
   },
   
   getSupportedFlavours: function()
   {
     var flavours = new FlavourSet();
+    flavours.appendFlavour("privatefragmentfile");
     flavours.appendFlavour("text/html");
     flavours.appendFlavour("text/unicode");
     return flavours;
